fix tree filling for default `Vector` constructor
When calling `subtree_fill_to_length` as part of a default constructor
for a vector that has a non-power-of-2 length, the computation of the
filler hash for unused nodes was incorrect.

e.g., the following would fail, and `hash_tree_root` would be incorrect.

```python
class Foo(Container):
    elems: Vector[Bytes32, 5]

assert Foo() == Foo(elems=[Bytes32()]*5)
```

Fixing the logic to use the correct `zero_node` for the empty subtree
addresses this issue; implicit and explicit zeroing now behave same.